Create new gimp-pyconsole module.
