+++ plug-ins/ipython-console/gimp/plugins/console/__init__.py
@@ -1,0 +1,7 @@
+import gi
+
+gi.require_version('Gimp', '3.0')
+gi.require_version('GimpUi', '3.0')
+gi.require_version('Gtk', '3.0')
+
+from .main import ConsoleDialog
