+++ plug-ins/ipython-console/gimp/plugins/console/pyconsole.py
